requirements.txt
pip==22.3.1
h5py==3.7.0
Cython==0.29.32
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.3.0
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
importlib_metadata==4.8.1
keyring==23.1.0
lockfile==0.12.2
lxml==4.6.3
mock==4.0.3
numpy==1.23.4
ordereddict==1.1
pandas==1.3.4
Pillow==9.3.0
protobuf==3.14.0
pyOpenSSL==22.1.0
simplejson==3.17.6
streamlit==1.11.0
toml==0.10.2
tornado==6.1
typing_extensions==3.10.0.2
zipp==3.6.0
imbalanced-learn==0.9.1
